Refactor Date Fields and Update FilterScreen to Use DateFields Widget
Refactor the `ReturnDateFields` widget by commenting it out and introduce a new `DateFields` widget. Update the `FilterScreen` to use the new `DateFields` widget for both departure and return date fields. Also, modify `FilterScreen` to be a `StatefulWidget` to manage `TextEditingController` instances and dispose of them properly. Update the `onPressed` method of the `CustomButton` to pass the correct filter parameters.